perf: admin tracking use server

+++ src/app/api/tracking/[trackingNumber]/route.js
@@ -1,7 +1,5 @@
 import { Tracking, TrackingItem } from '@/models'
 import { NextResponse } from 'next/server'
-import { revalidateTag } from 'next/cache'
-import { catching } from '@/config/const'
 
 export const GET = async (req, res) => {
   const { trackingNumber } = res.params
@@ -16,7 +14,6 @@ export const GET = async (req, res) => {
       }
     })
 
-    if (!res) return NextResponse.json('Tracking Number invalid!!!', { status: 404 })
     return NextResponse.json(res, { status: 200 })
   } catch (error) {
     return NextResponse.json({ error }, { status: 500 })
@@ -29,8 +26,6 @@ export const PATCH = async (req, res) => {
     const data = await req.json()
     const resTracking = await Tracking.update(data, { where: { id } })
 
-    if (!resTracking) return NextResponse.json('Tracking Id invalid!!!', { status: 404 })
-    revalidateTag(catching.GET_TRACKING_BY_ID)
     return NextResponse.json(resTracking, { status: 200 })
   } catch (error) {
     return NextResponse.json({ error }, { status: 500 })
@@ -40,8 +35,12 @@ export const PATCH = async (req, res) => {
 export const DELETE = async (req, res) => {
   const { trackingNumber: id } = res.params
   try {
-    await Tracking.destroy({ where: { id } })
-    revalidateTag(catching.GET_TRACKING_BY_ID)
+    const isExistTracking = await Tracking.findOne({ where: { id } })
+    if (!isExistTracking) throw new Error()
+
+    await TrackingItem.destroy({ where: { trackingId: id } }).then(async () =>
+      Tracking.destroy({ where: { id } })
+    )
     return NextResponse.json(id, { status: 200 })
   } catch (error) {
     return NextResponse.json({ error }, { status: 500 })
